requirements.txt
httpx
pydantic
openai
python-dotenv
termcolor